Swap failing CircleCI status for passing GH Action
I'm ambivalent about Circle vs. Actions, but the former is failing
missing a key in order to clone the repo, and I think I lack access to
fix that. Since it's only a simple workflow, seemed easiest to recreate
in Actions - and as that passes (as of
d62f1625be1f7d50439e3f0d5cbfdefcc2212bca) it seems to make more sense to
use that for a nice green status than red from Circle.